用户删除及分页查询功能提交

+++ src/main/java/com/worm/little/mapper/SysUserMapper.java
@@ -7,11 +7,11 @@ import java.util.List;
 
 @Repository
 public interface SysUserMapper {
-    int deleteByPrimaryKey(Integer userId);
+    int deleteByPrimaryKey(Long userId);
 
     int insert(SysUser record);
 
-    SysUser selectByPrimaryKey(Integer userId);
+    SysUser selectByPrimaryKey(Long userId);
 
     List<SysUser> selectAll();
